player.[h,c] + enemy can kill player
